requirements.txt
asgiref==3.8.1
Django==5.1.5
django-extensions==3.2.3
django-leaflet==0.31.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
django-filter~=24.3.0
djangorestframework-gis~=1.0.0
djangorestframework~=3.15.0